requirements.txt
streamlit==1.49.1
pandas==2.2.3
gspread==6.2.1
google-auth==2.40.3
reportlab==4.4.3